Fix joint trajectory constraint
This PR solves the issue that appears in ros2_control when the simulation loads.

parameter 'constraints.joint1.trajectory' has invalid type: expected [double] got [integer]

The command that I used to reproduce this was:

ros2 launch xarm_moveit_config xarm6_moveit_fake.launch.py

But the issue is valid for the 3 robots.

Signed-off-by: Emiliano Borghi <emiliano.borghi@gmail.com>